services/provenance-svc/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
Pillow
